Implement "ReplaceWhitespace" config option
Allows the user to specify a replacement character for any whitespace present in the search string.

This allows better configuration for sites that don't separate spaces with "+" (the default), such as Wikipedia (which uses an underscore "_").

By default, this change will set the inner default value of the new `ReplaceWhitespace` property to a space " ". This ensures backwards compatibility with any current setup or search patterns any user has.